+++ pyinstrument/renderers/__init__.py
@@ -1,0 +1,3 @@
+from pyinstrument.renderers.console import ConsoleRenderer
+from pyinstrument.renderers.html import HTMLRenderer
+from pyinstrument.renderers.json import JSONRenderer
