+++ pyinstrument/renderers/html_resources/profile.js
