+++ pyinstrument/renderers/html_resources/style.css
