+++ pyinstrument/renderers/html_resources/triangle_hide.png
