+++ pyinstrument/renderers/html_resources/triangle_show.png
